refactor: 파일 구조 변경(상품 관련 파일 상품 폴더로 모아놓음)
